refactor: split fetch preset logic from component
